# api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Cloud Storage Service API
  description: |
    # Cloud Storage Service

    基于Go语言和Gin框架构建的网盘文件存储服务，提供完备的RESTful API接口。

    ## 功能特性

    - 用户认证和授权
    - 文件上传和下载
    - 文件管理和操作
    - 文件夹管理
    - 文件分享
    - 回收站功能
    - 文件版本控制

    ## 认证

    使用Bearer Token进行认证。在登录成功后，将获得的访问令牌添加到请求头中：

    ```
    Authorization: Bearer <access_token>
    ```

  version: 1.0.0
  contact:
    name: Cloud Storage Team
    email: support@cloud-storage.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发服务器
  - url: https://api.cloud-storage.example.com/api/v1
    description: 生产服务器
...
paths:
...
  /files:
    get:
      tags: [文件]
      summary: 获取文件列表
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ParentID'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        '401':
          description: 未认证
        '500':
          description: 服务器内部错误

    post:
      tags: [文件, 目录]
      summary: 创建文件或目录
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          description: 请求参数错误
        '401':
          description: 未认证
        '409':
          description: 文件或目录已存在
        '500':
          description: 服务器内部错误
  /files/{id}:
    get:
      tags: [文件]
      summary: 获取文件信息
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileID'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '401':
          description: 未认证
        '403':
          description: 权限不足
        '404':
          description: 文件不存在
        '500':
          description: 服务器内部错误

    put:
      tags: [文件]
      summary: 更新文件信息
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          description: 请求参数错误
        '401':
          description: 未认证
        '403':
          description: 权限不足
        '404':
          description: 文件不存在
        '409':
          description: 文件名冲突
        '500':
          description: 服务器内部错误

    delete:
      tags: [文件]
      summary: 删除文件
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileID'
        - name: permanent
          in: query
          description: 是否永久删除
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 删除成功
        '401':
          description: 未认证
        '403':
          description: 权限不足
        '404':
          description: 文件不存在
        '500':
          description: 服务器内部错误
  /files/{id}/download:
    get:
      tags: [文件]
      summary: 下载文件
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileID'
      responses:
        '200':
          description: 文件下载
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: 未认证
        '403':
          description: 权限不足
        '404':
          description: 文件不存在
        '500':
          description: 服务器内部错误
  /upload:
    post:
      tags: [文件]
      summary: 上传文件
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 上传的文件
                parent_id:
                  type: string
                  format: uuid
                  description: 父目录ID
                is_public:
                  type: boolean
                  description: 是否公开
                  default: false
                override:
                  type: boolean
                  description: 是否覆盖同名文件
                  default: false
      responses:
        '201':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          description: 请求参数错误
        '401':
          description: 未认证
        '403':
          description: 存储配额不足
        '409':
          description: 文件已存在
        '413':
          description: 文件太大
        '500':
          description: 服务器内部错误

  # 系统相关
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    FileID:
      name: id
      in: path
      required: true
      description: 文件ID
      schema:
        type: string
        format: uuid
    ParentID:
      name: parent_id
      in: query
      required: false
      description: 父目录ID
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      description: 页码
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      description: 每页大小
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortBy:
      name: sort_by
      in: query
      required: false
      description: 排序字段
      schema:
        type: string
        enum: [name, size, created_at, updated_at]
        default: name
    SortOrder:
      name: sort_order
      in: query
      required: false
      description: 排序顺序
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  schemas:
...
    FileCreateRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          example: "document.pdf"
        parent_id:
          type: string
          format: uuid
          nullable: true
          example: "123e4567-e89b-12d3-a456-426614174000"
        type:
          type: string
          enum: [file, directory]
          example: "file"
        is_public:
          type: boolean
          default: false
          example: false
    FileUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: "renamed.pdf"
        parent_id:
          type: string
          format: uuid
          nullable: true
          example: "123e4567-e89b-12d3-a456-426614174000"
        is_public:
          type: boolean
          example: true
    FileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "document.pdf"
        path:
          type: string
          example: "/documents/document.pdf"
        size:
          type: integer
          example: 1048576
        mime_type:
          type: string
          example: "application/pdf"
        type:
          type: string
          enum: [file, directory]
          example: "file"
        is_public:
          type: boolean
          example: false
        share_token:
          type: string
          nullable: true
          example: "abc123def456ghi789"
        version:
          type: integer
          example: 1
        user_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        parent_id:
          type: string
          format: uuid
          nullable: true
          example: "123e4567-e89b-12d3-a456-426614174000"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
    FileListResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileResponse'
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20

    # 错误响应
...